feat: add wbtc token

+++ contracts/interfaces/IWrappedERC20.sol
@@ -4,7 +4,7 @@ pragma solidity ^0.8.17;
 import {IERC20} from "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 
 interface IWrappedERC20 is IERC20 {
-    function mint(address _to, uint _amount) external;
+    function mint(address _to, uint256 _amount) external;
 
-    function burn(address _from, uint _amount) external;
+    function burn(address _from, uint256 _amount) external;
 }
